2/13　Gimic_EmissionスクリプトをScriptフォルダに追加。　詳細はWikiに記載
