+++ PinkBall/Assets/Fungus/Integrations/Playmaker/Fungus-PlayMaker.unitypackage.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 0d79320315e49e54a8b41cc5ae74251c
-folderAsset: yes
-timeCreated: 1549896964
-licenseType: Free
+guid: 81dbcde0f90df4e9ba9ca2794490e57a
+timeCreated: 1491251885
+licenseType: Store
 DefaultImporter:
   externalObjects: {}
   userData:
