+++ PinkBall/Assets/Materials/test_hamaScene/Material/hama_GimikMaterial_Red.mat
@@ -72,5 +72,5 @@ Material:
     - _UVSec: 0
     - _ZWrite: 1
     m_Colors:
-    - _Color: {r: 0.47794116, g: 0.077314, b: 0.077314, a: 1}
+    - _Color: {r: 0.40000004, g: 0.1764706, b: 0.18431373, a: 1}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
