+++ PinkBall/Assets/Scripts/Gimic_Emission.cs.meta
